+++ packages/assets-controllers/src/token-prices-service/codefi-v2.ts
@@ -108,6 +108,8 @@ export const SUPPORTED_CURRENCIES = [
   'mxn',
   // Malaysian Ringgit
   'myr',
+  // Monad
+  'mon',
   // Nigerian Naira
   'ngn',
   // Norwegian Krone