+++ packages/network-controller/src/create-auto-managed-network-client.test.ts
@@ -35,6 +35,7 @@ describe('createAutoManagedNetworkClient', () => {
     describe(`given configuration for a ${networkClientConfiguration.type} network client`, () => {
       it('allows the network client configuration to be accessed', () => {
         const { configuration } = createAutoManagedNetworkClient({
+          networkClientId: 'some-network-client-id',
           networkClientConfiguration,
           getRpcServiceOptions: () => ({
             fetch,
@@ -51,6 +52,7 @@ describe('createAutoManagedNetworkClient', () => {
         // If unexpected requests occurred, then Nock would throw
         expect(() => {
           createAutoManagedNetworkClient({
+            networkClientId: 'some-network-client-id',
             networkClientConfiguration,
             getRpcServiceOptions: () => ({
               fetch,
@@ -64,6 +66,7 @@ describe('createAutoManagedNetworkClient', () => {
 
       it('returns a provider proxy that has the same interface as a provider', () => {
         const { provider } = createAutoManagedNetworkClient({
+          networkClientId: 'some-network-client-id',
           networkClientConfiguration,
           getRpcServiceOptions: () => ({
             fetch,
@@ -97,6 +100,7 @@ describe('createAutoManagedNetworkClient', () => {
           });
 
           const { provider } = createAutoManagedNetworkClient({
+            networkClientId: 'some-network-client-id',
             networkClientConfiguration,
             getRpcServiceOptions: () => ({
               fetch,
@@ -145,6 +149,7 @@ describe('createAutoManagedNetworkClient', () => {
           const messenger = buildNetworkControllerMessenger();
 
           const { provider } = createAutoManagedNetworkClient({
+            networkClientId: 'some-network-client-id',
             networkClientConfiguration,
             getRpcServiceOptions,
             getBlockTrackerOptions,
@@ -166,6 +171,7 @@ describe('createAutoManagedNetworkClient', () => {
           });
           expect(createNetworkClientMock).toHaveBeenCalledTimes(1);
           expect(createNetworkClientMock).toHaveBeenCalledWith({
+            id: 'some-network-client-id',
             configuration: networkClientConfiguration,
             getRpcServiceOptions,
             getBlockTrackerOptions,
@@ -204,6 +210,7 @@ describe('createAutoManagedNetworkClient', () => {
           const messenger = buildNetworkControllerMessenger();
 
           const autoManagedNetworkClient = createAutoManagedNetworkClient({
+            networkClientId: 'some-network-client-id',
             networkClientConfiguration,
             getRpcServiceOptions,
             getBlockTrackerOptions,
@@ -227,6 +234,7 @@ describe('createAutoManagedNetworkClient', () => {
           });
 
           expect(createNetworkClientMock).toHaveBeenNthCalledWith(1, {
+            id: 'some-network-client-id',
             configuration: networkClientConfiguration,
             getRpcServiceOptions,
             getBlockTrackerOptions,
@@ -234,6 +242,7 @@ describe('createAutoManagedNetworkClient', () => {
             isRpcFailoverEnabled: false,
           });
           expect(createNetworkClientMock).toHaveBeenNthCalledWith(2, {
+            id: 'some-network-client-id',
             configuration: networkClientConfiguration,
             getRpcServiceOptions,
             getBlockTrackerOptions,
@@ -272,6 +281,7 @@ describe('createAutoManagedNetworkClient', () => {
           const messenger = buildNetworkControllerMessenger();
 
           const autoManagedNetworkClient = createAutoManagedNetworkClient({
+            networkClientId: 'some-network-client-id',
             networkClientConfiguration,
             getRpcServiceOptions,
             getBlockTrackerOptions,
@@ -295,6 +305,7 @@ describe('createAutoManagedNetworkClient', () => {
           });
 
           expect(createNetworkClientMock).toHaveBeenNthCalledWith(1, {
+            id: 'some-network-client-id',
             configuration: networkClientConfiguration,
             getRpcServiceOptions,
             getBlockTrackerOptions,
@@ -302,6 +313,7 @@ describe('createAutoManagedNetworkClient', () => {
             isRpcFailoverEnabled: true,
           });
           expect(createNetworkClientMock).toHaveBeenNthCalledWith(2, {
+            id: 'some-network-client-id',
             configuration: networkClientConfiguration,
             getRpcServiceOptions,
             getBlockTrackerOptions,
@@ -313,6 +325,7 @@ describe('createAutoManagedNetworkClient', () => {
 
       it('returns a block tracker proxy that has the same interface as a block tracker', () => {
         const { blockTracker } = createAutoManagedNetworkClient({
+          networkClientId: 'some-network-client-id',
           networkClientConfiguration,
           getRpcServiceOptions: () => ({
             fetch,
@@ -372,6 +385,7 @@ describe('createAutoManagedNetworkClient', () => {
           });
 
           const { blockTracker } = createAutoManagedNetworkClient({
+            networkClientId: 'some-network-client-id',
             networkClientConfiguration,
             getRpcServiceOptions: () => ({
               fetch,
@@ -441,6 +455,7 @@ describe('createAutoManagedNetworkClient', () => {
           const messenger = buildNetworkControllerMessenger();
 
           const { blockTracker } = createAutoManagedNetworkClient({
+            networkClientId: 'some-network-client-id',
             networkClientConfiguration,
             getRpcServiceOptions,
             getBlockTrackerOptions,
@@ -458,6 +473,7 @@ describe('createAutoManagedNetworkClient', () => {
           await blockTracker.checkForLatestBlock();
           expect(createNetworkClientMock).toHaveBeenCalledTimes(1);
           expect(createNetworkClientMock).toHaveBeenCalledWith({
+            id: 'some-network-client-id',
             configuration: networkClientConfiguration,
             getRpcServiceOptions,
             getBlockTrackerOptions,
@@ -496,6 +512,7 @@ describe('createAutoManagedNetworkClient', () => {
           const messenger = buildNetworkControllerMessenger();
 
           const autoManagedNetworkClient = createAutoManagedNetworkClient({
+            networkClientId: 'some-network-client-id',
             networkClientConfiguration,
             getRpcServiceOptions,
             getBlockTrackerOptions,
@@ -513,6 +530,7 @@ describe('createAutoManagedNetworkClient', () => {
           });
 
           expect(createNetworkClientMock).toHaveBeenNthCalledWith(1, {
+            id: 'some-network-client-id',
             configuration: networkClientConfiguration,
             getRpcServiceOptions,
             getBlockTrackerOptions,
@@ -520,6 +538,7 @@ describe('createAutoManagedNetworkClient', () => {
             isRpcFailoverEnabled: false,
           });
           expect(createNetworkClientMock).toHaveBeenNthCalledWith(2, {
+            id: 'some-network-client-id',
             configuration: networkClientConfiguration,
             getRpcServiceOptions,
             getBlockTrackerOptions,
@@ -558,6 +577,7 @@ describe('createAutoManagedNetworkClient', () => {
           const messenger = buildNetworkControllerMessenger();
 
           const autoManagedNetworkClient = createAutoManagedNetworkClient({
+            networkClientId: 'some-network-client-id',
             networkClientConfiguration,
             getRpcServiceOptions,
             getBlockTrackerOptions,
@@ -575,6 +595,7 @@ describe('createAutoManagedNetworkClient', () => {
           });
 
           expect(createNetworkClientMock).toHaveBeenNthCalledWith(1, {
+            id: 'some-network-client-id',
             configuration: networkClientConfiguration,
             getRpcServiceOptions,
             getBlockTrackerOptions,
@@ -582,6 +603,7 @@ describe('createAutoManagedNetworkClient', () => {
             isRpcFailoverEnabled: true,
           });
           expect(createNetworkClientMock).toHaveBeenNthCalledWith(2, {
+            id: 'some-network-client-id',
             configuration: networkClientConfiguration,
             getRpcServiceOptions,
             getBlockTrackerOptions,
@@ -608,6 +630,7 @@ describe('createAutoManagedNetworkClient', () => {
         ],
       });
       const { blockTracker, destroy } = createAutoManagedNetworkClient({
+        networkClientId: 'some-network-client-id',
         networkClientConfiguration,
         getRpcServiceOptions: () => ({
           fetch,